Add more R1 runtime values, attempt to allow real-time editing and freezing of values

+++ Assets/Scripts/DataTypes/Rayman1/Common/Global/R1_RayMode.cs
@@ -1,6 +1,6 @@
 ﻿namespace R1Engine
 {
-    public enum R1_RayMode : sbyte
+    public enum R1_RayMode : short
     {
         PlaceRay = -1, // All other values are PlaceRay, but the game uses -1
         Rayman = 1,
